One .pre_commit_config.yaml example per test
